% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov2.R
\name{cov2}
\alias{cov2}
\title{Sample Covariance for 2 Vectors}
\usage{
cov2(x, y, integer = FALSE)
}
\arguments{
\item{x, y}{Integer or numeric vector.}

\item{integer}{Logical value for whether \code{x} and \code{y} are integer 
vectors.}
}
\value{
Numeric value.
}
\description{
This is a faster version of base R function \code{\link[stats]{cov}}. For 
optimal speed, use \code{integer = TRUE} if \code{x} and \code{y} are integer 
vectors and \code{integer = FALSE} otherwise. This function uses a one-pass 
covariance calculation that makes it susceptible to precision issues if 
\code{x} values are very large.
}
\examples{
# cov2 vs. cov for integer vector
x <- rpois(1000, lambda = 5)
y <- rpois(1000, lambda = 5)
all.equal(cov(x, y), cov2(x, y, integer = TRUE))
benchmark(cov(x, y), cov2(x, y, integer = TRUE), replications = 5000)

# cov2 vs. cov for numeric vector
x <- rnorm(1000)
y <- rnorm(1000)
all.equal(cov(x, y), cov2(x, y))
benchmark(cov(x, y), cov2(x, y), replications = 5000)


}
