% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff2.R
\name{diff2}
\alias{diff2}
\title{Lagged Differences}
\usage{
diff2(x, lag = 1, integer = FALSE)
}
\arguments{
\item{x}{Integer or numeric vector.}

\item{lag}{Integer value.}

\item{integer}{Logical value for whether \code{x} is an integer vector.}
}
\value{
Integer or numeric value.
}
\description{
This is a faster version of base R function \code{\link[base]{diff}}. For 
optimal speed, use \code{integer = TRUE} if \code{x} is an integer vector and 
\code{integer = FALSE} otherwise.
}
\examples{
# diff2 vs. diff for integer vector and 1-unit lag
x <- rpois(1000, lambda = 5)
all.equal(diff(x), diff2(x, integer = TRUE))
benchmark(diff(x), diff2(x, integer = TRUE), replications = 5000)

# diff2 vs. diff for integer vector and 3-unit lag
x <- rpois(1000, lambda = 5)
all.equal(diff(x, 2), diff2(x, 2, integer = TRUE))
benchmark(diff(x, 2), diff2(x, 2, integer = TRUE), replications = 5000)

# diff2 vs. diff for numeric vector and 1-unit lag
x <- rnorm(1000)
all.equal(diff(x), diff2(x))
benchmark(diff(x), diff2(x), replications = 5000)

# diff2 vs. diff for numeric vector and 3-unit lag
x <- rnorm(1000)
all.equal(diff(x, 2), diff2(x, 2))
benchmark(diff(x, 2), diff2(x, 2), replications = 5000)


}
