% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean2.R
\name{mean2}
\alias{mean2}
\title{Arithmetic Mean}
\usage{
mean2(x, integer = FALSE)
}
\arguments{
\item{x}{Integer or numeric vector.}

\item{integer}{Logical value for whether \code{x} is an integer vector.}
}
\value{
Numeric value.
}
\description{
Faster version of base R function \code{\link[base]{mean}}. For optimal speed, 
use \code{Integer = TRUE} if \code{x} is an integer vector and 
\code{integer = FALSE} otherwise.
}
\examples{
# mean2 vs. mean for integer vector
x <- rpois(1000, lambda = 5)
all.equal(mean(x), mean2(x, TRUE))
benchmark(mean(x), mean2(x, TRUE), replications = 1000)

# mean2 vs. mean for numeric vector
x <- rnorm(1000)
all.equal(mean(x), mean2(x))
benchmark(mean(x), mean2(x), replications = 1000)


}
