% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle2.R
\name{rle2}
\alias{rle2}
\title{Run Length Encoding}
\usage{
rle2(x, integer = FALSE)
}
\arguments{
\item{x}{Integer, numeric, or logical vector.}

\item{integer}{Logical value for whether \code{x} is an integer vector.}
}
\value{
Integer or numeric matrix.
}
\description{
This is a faster version of base R function \code{\link[base]{rle}}. Returns 
matrix where first column is value, second column is length. For optimal 
speed, use \code{integer = TRUE} if \code{x} is an integer vector and 
\code{integer = FALSE} otherwise. Compatible with logical vectors (use 
\code{integer = TRUE}) but replaces TRUE/FALSE with 1/0; not compatible with 
character vectors.
}
