% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd2.R
\name{sd2}
\alias{sd2}
\title{Sample Standard Deviation}
\usage{
sd2(x, integer = FALSE)
}
\arguments{
\item{x}{Integer or numeric vector.}

\item{integer}{Logical value for whether \code{x} is an integer vector.}
}
\value{
Numeric value.
}
\description{
Faster version of base R function \code{\link[stats]{sd}}. For optimal speed, 
use \code{Integer = TRUE} if \code{x} is an integer vector and 
\code{integer = FALSE} otherwise.
}
\examples{
# sd2 vs. sd for integer vector
x <- rpois(1000, lambda = 5)
all.equal(sd(x), sd2(x, integer = TRUE))
benchmark(sd(x), sd2(x, integer = TRUE), replications = 2000)

# sd2 vs. sd for numeric vector
x <- rnorm(1000)
all.equal(sd(x), sd2(x))
benchmark(sd(x), sd2(x), replications = 2000) 


}
