% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var2.R
\name{var2}
\alias{var2}
\title{Sample Variance}
\usage{
var2(x, integer = FALSE)
}
\arguments{
\item{x}{Integer or numeric vector.}

\item{integer}{Logical value for whether \code{x} is an integer vector.}
}
\value{
Numeric value.
}
\description{
Faster version of base R function \code{\link[stats]{var}}. For optimal speed, 
use \code{Integer = TRUE} if \code{x} is an integer vector and 
\code{integer = FALSE} otherwise.
}
\examples{
# var2 vs. var for integer vector
x <- rpois(1000, lambda = 5)
all.equal(var(x), var2(x, TRUE))
benchmark(var(x), var2(x, TRUE), replications = 1000)

# var2 vs. var for numeric vector
x <- rnorm(1000)
all.equal(var(x), var2(x))
benchmark(var(x), var2(x), replications = 1000)


}
