% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_mean2.R
\name{weighted.mean2}
\alias{weighted.mean2}
\title{Weighted Arithmetic Mean}
\usage{
weighted.mean2(x, w, x.integer = FALSE, w.integer = FALSE)
}
\arguments{
\item{x}{Integer or numeric vector.}

\item{w}{Integer or numeric vector of weights.}

\item{x.integer}{Logical value for whether \code{x} is an integer vector.}

\item{w.integer}{Logical value for whether \code{w} is an integer vector.}
}
\value{
Numeric value.
}
\description{
This is a faster version of base R function 
\code{\link[stats]{weighted.mean}}. For optimal speed, specify 
\code{x.integer = TRUE} and \code{w.integer = TRUE} if \code{x} and \code{w} 
are integer vectors.
}
\examples{
# weighted.mean2 vs. weighted.mean for integer vectors
x <- rpois(1000, lambda = 5)
w <- rpois(1000, lambda = 1)
all.equal(weighted.mean(x, w), weighted.mean2(x, w, TRUE, TRUE))
benchmark(weighted.mean(x, w), weighted.mean2(x, w, TRUE, TRUE), 
          replications = 1000)

# weighted.mean2 vs. weighted.mean for numeric vectors
x <- rnorm(1000)
w <- rexp(1000) 
all.equal(weighted.mean(x, w), weighted.mean2(x, w))
benchmark(weighted.mean(x), weighted.mean2(x, w), 
          replications = 1000)


}
