% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean2.R
\name{mean2}
\alias{mean2}
\title{Arithmetic Mean}
\usage{
mean2(x, integer = FALSE)
}
\arguments{
\item{x}{Integer or numeric vector.}
}
\value{
Numeric value.
}
\description{
This is a faster version of the base R function \code{\link[base]{mean}}, 
defined simply as \code{sum(x) / length(x)}. Speed advantage diminishes with 
\code{length(x)}.
}
\examples{
# mean2 vs. mean for integer vector
x <- rpois(1000, lambda = 5)
all.equal(mean(x), mean2(x))
benchmark(mean(x), mean2(x), replications = 1000)

# mean2 vs. mean for numeric vector
x <- rnorm(1000)
all.equal(mean(x), mean2(x))
benchmark(mean(x), mean2(x), replications = 1000)


}
