% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean2.R
\name{mean2}
\alias{mean2}
\title{Arithmetic Mean}
\usage{
mean2(x, integer = FALSE)
}
\arguments{
\item{x}{Integer or numeric vector.}

\item{integer}{Logical value for whether \code{x} is an integer vector.}
}
\value{
Numeric value.
}
\description{
This is a faster version of the base R function \code{\link[base]{mean}}, 
defined simply as \code{sum(x) / length(x)}. The speed advantage diminishes 
with \code{length(x)}. Note that a C++ implementation is faster for very 
large vectors (i.e. length around 25,000), but much slower for shorter 
vectors, so it is not currently utilized.
}
\examples{
# mean2 vs. mean for integer vector
x <- rpois(1000, lambda = 5)
all.equal(mean(x), mean2(x))
benchmark(mean(x), mean2(x), replications = 5000)

# mean2 vs. mean for numeric vector
x <- rnorm(1000)
all.equal(mean(x), mean2(x))
benchmark(mean(x), mean2(x), replications = 5000)


}
